Move WebConsole::Template::Context into WebConsole::View
